Add minimal test framework including PoC tests
This invovles bumping the minSdkVersion 8->9, and correcting the
defaultConfig applicationId in build.gradle.

Several new dependencies were added, along with whole new directory
structures containing the 'proof of concept' example tests.

This represents a significant step forward in the development of this
app, in particular in how we go about writing code.

Also included are some XML config files in the .idea dir. Obviously
necessary is the runConfiguration.xml file, which is part of running
the tests. Less necessary - but included for the sake of completeness -
are misc.xml and the .idea/codeStyles/ dir, which contains
codeStyleConfig.xml and Project.xml.