Refactor: moved unit action code to mechanics pkg
Both the moved methods are static, which I'm not sure about, and would
welcome input on.

I've written several new methods in the Unit.java class that turn it
inot nothing more than a data store about the unit. There is nothing in
there but fancy getters and setters.

Incidentally I was thinking about adding a unit range variable, but
decided that units that move >1 sector per turn aren't necessary for the
MVP.