Adds the concept of players to the game
Players now exist, having the attributes of faction, intelligence (human
or cpu player), extant (alive, dead, hiberating, outside the galaxy,
etc), and credits.

Credits is potentially a placeholder for a more complex sort of
inter-species economic system, but for now it can stay as an int.

Extant has two extra values besides alive and dead, which are not MVP
and can be taken out if preferred. They're more there for this weekends
discussions.