Adding different types of ships
This only adds two types of ships: CombatShip and ColonyShip.

They both extend from Spaceship, which is the abstract parent class of
all mobile Spacecraft, which is the interface for all spacebourne
constructions of all factions.

It also adds the Enum SpacecraftOrders, which all Spacecraft will be
able to return a list of, to inform the context menu of which orders it
can be given.

Closes #64