Systems are now ownable
Systems now have an 'owner' field, of type Faction. This assumes there
will only be one instance of each faction in the game.

Also added a static function in System.java that creates systems and
puts them in the galaxy, purely as a way of only having to specify the
xy coordinates once, as that is error prone (or I'm error prone when
creating systems).

Issue #62