Bug: two ships in same sector, one is stuck
When two ships in the same sector are ordered to move, one will be left
behind due to a bug in how the list of ships is processed.

The fix is that a reverse ordered list of indicies of ships in the
sector that leave the sector is removed from the list of ships in the
sector _after_ the orders of ships in that sector are executed, so that
the removal of items doesn't interfere with either the execution of
orders or the removal of later ships.

This also adds some debug logging and removes an unnecessary parameter
passing from an unrelated method.

Fixes #92